[droidian-apt-config] Ship every repository required to run Droidian
Previously, setting up repositories was the job of whatever bootstraps the
Droidian image.

This means that the repositories themselves weren't tracked in a package
and thus need to be modified manually in case of distribution upgrades.

This might be desirable by some people, and definitely handy on desktop
machines, but it falls outside the expectations of the upgrade flow
of a modern mobile OS.

Let's ship everything needed for Droidian in this package. Upgrading
this package allows to ship an eventual upgrade without manual intervention
beyond the standard upgrade tools.

The only drawback is that mirrors are now out of the question, but:

a) only the Debian repository has mirrors

b) due to the power of default, most people stick with deb.debian.org since
   it's what we ship with (as opposed as installing Debian via d-i, where
   the mirror to use is asked during installation)

c) deb.debian.org is geographically distributed and backed by a reliable CDN

Perhaps in future only the debian.sources file can be installed as a conffile,
to allow changing it (note: the file can still be diverted and replaced with
a custom one)

The old so-called "extrepo" repositories are removed only during upgrades on
an old system. It's a bug of the bootstrap tool if extrepo files are kept after
bootstrap.

/etc/apt/sources.list is not touched for obvious reasons. Bootstrap tools are
required to keep that file empty.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>