[sources.list.d] phone: pick squeekboard from droidian/squeekboard@upgrade-1-24-0
squeekboard has been removed from testing, and the version from sid is not
installable.

phone.list is not the right place for this, but as it's something that should
be really temporary (once the t64 migration finishes, either it gets into
trixie again or the version from sid becomes installable) let's not overdo
it.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>